working on parsing [not done]

+++ src/atom.h
@@ -3,11 +3,11 @@
 class Atom {
     public:
         float coeff;
-        int exp;
+        float exp;
         char var;
 
         Atom() : coeff(0), exp(0), var('x') {}
-        Atom(float a, char v, int e) : coeff(a), exp(e), var(v) {}
+        Atom(float a, char v, float e) : coeff(a), exp(e), var(v) {}
 
         float eval(float x) {
             return coeff * std::pow(x, exp);
